[OPIK-126] [UX Improvements] Add label for experiments that use a deleted dataset
