+++ apps/opik-frontend/src/components/pages/CompareExperimentsPage/CompareExperimentsDetails.tsx
@@ -1,27 +1,26 @@
 import React, { useEffect, useMemo } from "react";
 import sortBy from "lodash/sortBy";
+import uniq from "lodash/uniq";
+import isUndefined from "lodash/isUndefined";
+import { BooleanParam, useQueryParam } from "use-query-params";
 import {
-  ArrowUpRight,
   Clock,
-  Database,
   FlaskConical,
   Maximize2,
   Minimize2,
   PenLine,
 } from "lucide-react";
 
-import useAppStore from "@/store/AppStore";
 import useBreadcrumbsStore from "@/store/BreadcrumbsStore";
 import FeedbackScoreTag from "@/components/shared/FeedbackScoreTag/FeedbackScoreTag";
 import { Experiment } from "@/types/datasets";
 import { TableBody, TableCell, TableRow } from "@/components/ui/table";
 import { Tag } from "@/components/ui/tag";
-import { Link } from "@tanstack/react-router";
 import { formatDate } from "@/lib/date";
-import uniq from "lodash/uniq";
-import isUndefined from "lodash/isUndefined";
 import { Button } from "@/components/ui/button";
-import { BooleanParam, useQueryParam } from "use-query-params";
+import ResourceLink, {
+  RESOURCE_TYPE,
+} from "@/components/shared/ResourceLink/ResourceLink";
 
 type CompareExperimentsDetailsProps = {
   experimentsIds: string[];
@@ -31,7 +30,6 @@ type CompareExperimentsDetailsProps = {
 const CompareExperimentsDetails: React.FunctionComponent<
   CompareExperimentsDetailsProps
 > = ({ experiments, experimentsIds }) => {
-  const workspaceName = useAppStore((state) => state.activeWorkspaceName);
   const setBreadcrumbParam = useBreadcrumbsStore((state) => state.setParam);
 
   const isCompare = experimentsIds.length > 1;
@@ -208,17 +206,11 @@ const CompareExperimentsDetails: React.FunctionComponent<
             <div className="truncate">{formatDate(experiment?.created_at)}</div>
           </Tag>
         )}
-        <Link
-          to={"/$workspaceName/datasets/$datasetId/items"}
-          params={{ workspaceName, datasetId: experiment?.dataset_id }}
-          className="max-w-full"
-        >
-          <Tag size="lg" variant="gray" className="flex items-center gap-2">
-            <Database className="size-4 shrink-0" />
-            <div className="truncate">{experiment?.dataset_name}</div>
-            <ArrowUpRight className="size-4 shrink-0" />
-          </Tag>
-        </Link>
+        <ResourceLink
+          id={experiment?.dataset_id}
+          name={experiment?.dataset_name}
+          resource={RESOURCE_TYPE.dataset}
+        />
       </div>
       {renderSubSection()}
       {renderCompareFeedbackScores()}
