+++ apps/opik-frontend/src/components/shared/DataTableCells/ResourceCell.tsx
@@ -1,25 +1,12 @@
 import React from "react";
 import get from "lodash/get";
 import { CellContext } from "@tanstack/react-table";
-import { Link } from "@tanstack/react-router";
-import { ArrowUpRight, Database } from "lucide-react";
 
 import CellWrapper from "@/components/shared/DataTableCells/CellWrapper";
-import TooltipWrapper from "@/components/shared/TooltipWrapper/TooltipWrapper";
-import useAppStore from "@/store/AppStore";
-import { Tag } from "@/components/ui/tag";
 
-export enum RESOURCE_TYPE {
-  dataset,
-}
-
-const RESOURCE_MAP = {
-  [RESOURCE_TYPE.dataset]: {
-    url: "/$workspaceName/datasets/$datasetId/items",
-    icon: Database,
-    param: "datasetId",
-  },
-};
+import ResourceLink, {
+  RESOURCE_TYPE,
+} from "@/components/shared/ResourceLink/ResourceLink";
 
 type CustomMeta = {
   nameKey?: string;
@@ -28,7 +15,6 @@ type CustomMeta = {
 };
 
 const ResourceCell = (context: CellContext<unknown, string>) => {
-  const workspaceName = useAppStore((state) => state.activeWorkspaceName);
   const { custom } = context.column.columnDef.meta ?? {};
   const cellData = context.row.original;
   const {
@@ -40,33 +26,14 @@ const ResourceCell = (context: CellContext<unknown, string>) => {
   const name = get(cellData, nameKey, undefined);
   const id = get(cellData, idKey, undefined);
 
-  if (!name || !id) return null;
-  const props = RESOURCE_MAP[resource];
-  const params: Record<string, string> = {
-    workspaceName,
-  };
-
-  params[props.param] = id;
+  if (!id) return null;
 
   return (
     <CellWrapper
       metadata={context.column.columnDef.meta}
       tableMetadata={context.table.options.meta}
     >
-      <Link
-        to={props.url}
-        params={params}
-        onClick={(event) => event.stopPropagation()}
-        className="max-w-full"
-      >
-        <TooltipWrapper content={name}>
-          <Tag size="lg" variant="gray" className="flex items-center gap-2">
-            <props.icon className="size-4 shrink-0" />
-            <div className="truncate">{name}</div>
-            <ArrowUpRight className="size-4 shrink-0" />
-          </Tag>
-        </TooltipWrapper>
-      </Link>
+      <ResourceLink id={id} name={name} resource={resource} />
     </CellWrapper>
   );
 };
